requirements.txt
Django==1.8.5
django-appconf==1.0.1
django-cors-headers==1.1.0
django-filter==0.11.0
django-flat-theme==1.1.1
django-geoposition==0.2.2
django-rest-swagger==0.3.4
djangorestframework==3.2.4
Markdown==2.6.2
Pillow==2.9.0
psycopg2==2.6.1
PyYAML==3.11
six==1.9.0
sorl-thumbnail==12.3
urllib3==1.12
